@@ -3,11 +3,15 @@
   "description": "",
   "version": "1.0.0",
   "dependencies": {
-    "@chakra-ui/react": "^1.8.9",
-    "@cmsgov/design-system": "^5.0.2",
+    "@chakra-ui/react": "2.5.5",
+    "@chakra-ui/toast": "2.0.0",
+    "@chakra-ui/styled-system": "^2",
+    "@cmsgov/design-system": "^11.1.0",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@hookform/resolvers": "^2.9.11",
+    "@tanstack/react-query": "^5",
+    "@tanstack/react-query-devtools": "^5",
     "@vitejs/plugin-react": "^4.3.1",
     "aws-amplify": "^6.6.4",
     "base-64": "^1.0.0",
@@ -20,13 +24,12 @@
     "node-fetch": "3.3.2",
     "node-forge": "1.3.1",
     "object-path": "^0.11.8",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18",
+    "react-dom": "^18",
     "react-dropzone": "^11.7.1",
     "react-hook-form": "^7.53.0",
     "react-icons": "^4.12.0",
     "react-idle-timer": "^4.6.4",
-    "react-query": "^3.39.3",
     "react-router-dom": "6.26.1",
     "react-textarea-autosize": "^8.5.3",
     "sass": "^1.80.4",
@@ -40,8 +43,8 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "coverage": "jest --coverage",
-    "test": "jest --setupFiles dotenv/config",
-    "watchTest": "jest --setupFiles dotenv/config --watch",
+    "test": "jest",
+    "watchTest": "jest --watch",
     "postinstall": "patch-package"
   },
   "jest": {
@@ -129,15 +132,15 @@
     "@babel/preset-env": "^7.13.15",
     "@babel/preset-typescript": "^7.16.0",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^12.1.0",
+    "@testing-library/react": "13.0.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/base-64": "^1.0.0",
     "@types/file-saver": "^2.0.5",
     "@types/jest": "^27.4.0",
     "@types/jest-axe": "^3.5.4",
     "@types/object-path": "^0.11.1",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@types/react-router-dom": "^5.3.2",
     "@types/uuid": "^8.3.3",
     "babel": "^6.23.0",
@@ -151,7 +154,7 @@
     "serverless-cloudfront-invalidate": "^1.5.0",
     "serverless-s3-sync": "^1.14.4",
     "ts-jest": "^27.1.5",
-    "typescript": "^4.4.4"
+    "typescript": "^5"
   },
   "resolutions": {
     "node-forge": "^1.3.1",
