@@ -34,13 +34,13 @@ export const TextField = ({
       label={label || ""}
       placeholder={placeholder}
       onChange={(e) => onChangeHandler(e)}
-      errorMessage={errorMessage}
+      errorMessage={errorMessage?.toString()}
       value={displayValue}
       {...props}
     />
   );
 };
-interface Props {
+interface Props extends React.PropsWithChildren {
   name: string;
   label?: string;
   placeholder?: string;
