@@ -34,7 +34,7 @@ export const ContainedButton = ({
         leftIcon={printIcon}
         rightIcon={plusIcon}
         onClick={onClick}
-        disabled={disabledStatus}
+        isDisabled={disabledStatus}
         borderRadius="sm"
         zIndex={zIndex}
         fontSize="1.2rem"
