@@ -1,11 +1,12 @@
-import { Box, Text, useToast } from "@chakra-ui/react";
+import { createStandaloneToast } from "@chakra-ui/toast";
+import { Box, Text } from "@chakra-ui/react";
 import { ContainedButton } from "components/ContainedButton";
 import { CoreSetAbbr } from "types";
 import { CoreSetTableItem } from "components/Table/types";
 import { useEditCoreSet } from "hooks/api";
 import { useParams } from "react-router-dom";
 import { useUser } from "hooks/authHooks";
-import { useQueryClient } from "react-query";
+import { useQueryClient } from "@tanstack/react-query";
 import { featuresByYear } from "utils/featuresByYear";
 
 interface Props {
@@ -63,11 +64,11 @@ export const SubmitCoreSetButton = ({
   const helperText = `Complete all ${helperTextFiller()}Core Set Questions${subSetTextFiller()}and ${helperTextFiller()}Core Set Measures${subSetTextFiller()}to submit ${
     featuresByYear.displayFFYLanguage ? "FFY" : ""
   } ${year}`;
-  const { mutate, isLoading } = useEditCoreSet();
+  const { mutate, isPending } = useEditCoreSet();
   const queryClient = useQueryClient();
   const userInfo = useUser();
 
-  const toast = useToast();
+  const { toast } = createStandaloneToast();
 
   const urlParams = useParams();
   const state = urlParams.state !== undefined ? urlParams.state : "";
@@ -78,7 +79,7 @@ export const SubmitCoreSetButton = ({
         <ContainedButton
           buttonText={"Submit Core Set"}
           disabledStatus={
-            isLoading ||
+            isPending ||
             !userInfo.isStateUser ||
             coreSetStatus !== CoreSetTableItem.Status.COMPLETED
           }
