@@ -1,4 +1,4 @@
-import { useMutation } from "react-query";
+import { useMutation } from "@tanstack/react-query";
 import { createMeasure } from "libs/api";
 import { CoreSetAbbr, Params } from "types";
 
@@ -31,5 +31,7 @@ const addMeasure = ({
 };
 
 export const useAddMeasure = () => {
-  return useMutation((data: CreateMeasure) => addMeasure({ ...data }));
+  return useMutation({
+    mutationFn: (data: CreateMeasure) => addMeasure({ ...data }),
+  });
 };
