@@ -1,13 +1,17 @@
-import { useMutation, useQuery } from "react-query";
+import { useMutation, useQuery } from "@tanstack/react-query";
 import { AdminBannerData } from "types";
 import { getBanner, writeBanner, deleteBanner } from "libs/api";
 
 export const useDeleteBanner = () => {
-  return useMutation((bannerKey: string) => deleteBanner(bannerKey));
+  return useMutation({
+    mutationFn: (bannerKey: string) => deleteBanner(bannerKey),
+  });
 };
 
 export const useWriteBanner = () => {
-  return useMutation((bannerData: AdminBannerData) => writeBanner(bannerData));
+  return useMutation({
+    mutationFn: (bannerData: AdminBannerData) => writeBanner(bannerData),
+  });
 };
 
 const _getBanner = async (bannerKey: string) => {
@@ -16,5 +20,8 @@ const _getBanner = async (bannerKey: string) => {
 };
 
 export const useGetBanner = (bannerKey: string) => {
-  return useQuery(bannerKey, () => _getBanner(bannerKey));
+  return useQuery({
+    queryKey: [bannerKey],
+    queryFn: () => _getBanner(bannerKey),
+  });
 };
