@@ -1,4 +1,4 @@
-import { useMutation } from "react-query";
+import { useMutation } from "@tanstack/react-query";
 import * as Api from "libs/api";
 import { CoreSetAbbr } from "types";
 
@@ -17,5 +17,7 @@ const deleteCoreSet = async ({ state, year, coreSet }: DeleteCoreSet) => {
 };
 
 export const useDeleteCoreSet = () => {
-  return useMutation((data: DeleteCoreSet) => deleteCoreSet(data));
+  return useMutation({
+    mutationFn: (data: DeleteCoreSet) => deleteCoreSet(data),
+  });
 };
