@@ -1,4 +1,4 @@
-import { useQuery } from "react-query";
+import { useQuery } from "@tanstack/react-query";
 import { getMeasureListInfo } from "libs/api";
 
 const getList = async () => {
@@ -6,7 +6,9 @@ const getList = async () => {
 };
 
 export const useGetMeasureListInfo = () => {
-  return useQuery(["measureListInfo"], () => getList(), {
+  return useQuery({
+    queryKey: ["measureListInfo"],
+    queryFn: getList,
     refetchOnWindowFocus: false,
   });
 };
