@@ -1,4 +1,4 @@
-import { useQuery } from "react-query";
+import { useQuery } from "@tanstack/react-query";
 import { getReportingYears } from "libs/api";
 import { usePathParams } from "./usePathParams";
 import { useParams } from "react-router-dom";
@@ -20,12 +20,14 @@ export const useGetReportingYears = () => {
   const { state, year } = useParams();
 
   if ((state || statePath) && (year || yearPath)) {
-    return useQuery(["reportingYears", state, year], () =>
-      getYears({
-        state: state || statePath,
-        year: year || yearPath,
-      })
-    );
+    return useQuery({
+      queryKey: ["reportingYears", state, year],
+      queryFn: () =>
+        getYears({
+          state: state || statePath,
+          year: year || yearPath,
+        }),
+    });
   }
   throw Error("state or year unavailable");
 };
