@@ -1,7 +1,7 @@
 import { CoreSet } from "./index";
 import { render, screen } from "@testing-library/react";
 import { RouterWrappedComp } from "utils/testing";
-import { QueryClient, QueryClientProvider } from "react-query";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { useApiMock } from "utils/testUtils/useApiMock";
 const queryClient = new QueryClient();
 
